+++ src/irc/irccontroller.cpp
@@ -114,6 +114,12 @@ IrcController::IrcController(QObject *parent)
     });
     connect(&m_console, &IrcStatusConsole::alertsChanged, this,
             &IrcController::statusChanged);
+    m_reducer.setConversationLog(&m_transcripts);
+}
+
+void IrcController::setTranscriptRoot(const QString &root)
+{
+    m_transcripts.setRoot(root);
 }
 
 IrcSession *IrcController::addSession(const IrcSessionConfig& config,