+++ src/irc/irccontroller.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "conversationlistmodel.h"
+#include "ircconversationlog.h"
 #include "irceventreducer.h"
 #include "irchighlight.h"
 #include "ircignore.h"
@@ -50,6 +51,7 @@ class IrcController : public QObject
 public:
     explicit IrcController(QObject *parent = nullptr);
 
+    void setTranscriptRoot(const QString &root);
     IrcSession *addSession(const IrcSessionConfig& config,
                            IrcTransport *transport,
                            IrcReconnectTimer *reconnectTimer = nullptr);
@@ -219,6 +221,7 @@ private:
 
     IrcSessionManager m_sessions;
     IrcStatusConsole m_console;
+    IrcConversationLog m_transcripts;
     IrcEventReducer m_reducer;
     IrcIgnoreStore m_ignores;
     IrcMuteStore m_mutes;